updated starups +small csv

+++ processed/startup-small.sql
@@ -4,7 +4,7 @@ CREATE TABLE Country (
   name varchar(70),
   region varchar(50),
   psychiatrist_count float,
-  population float,
+  population int,
   PRIMARY KEY(name)
 );
 CREATE TABLE Suicide_Rates (
